+++ internal/provider/provider_test.go
@@ -13,6 +13,24 @@ func protoV5ProviderFactories() map[string]func() (tfprotov5.ProviderServer, err
 	}
 }
 
+func providerVersion221() map[string]resource.ExternalProvider {
+	return map[string]resource.ExternalProvider{
+		"random": {
+			VersionConstraint: "2.2.1",
+			Source:            "hashicorp/random",
+		},
+	}
+}
+
+func providerVersion320() map[string]resource.ExternalProvider {
+	return map[string]resource.ExternalProvider{
+		"random": {
+			VersionConstraint: "3.2.0",
+			Source:            "hashicorp/random",
+		},
+	}
+}
+
 func providerVersion332() map[string]resource.ExternalProvider {
 	return map[string]resource.ExternalProvider{
 		"random": {